wip on meetings/videoconferences

+++ app/Interfaces/Implementations/LocalMediaAdapter.php
@@ -127,7 +127,7 @@ class LocalMediaAdapter implements MediaInterface
                 case 'Kanban':
                     $class = 'App\\'.$params['model'];
                     $model = (new $class)::where('id',$params['model_id'] )->get()->first();
-               
+
                     if ($model->isAccessible()){
                         return ($medium->mime_type != 'url') ? response()->file($path) : redirect($medium->path); //return file or url
                     }
@@ -224,7 +224,10 @@ class LocalMediaAdapter implements MediaInterface
         }
 
         if ($medium->subscriptions()->count() <= 1) {
-            Storage::disk(config('filesystems.default'))->delete($medium->path.$medium->medium_name);
+            if ($medium->mime_type != 'url')
+            {
+                Storage::disk(config('filesystems.default'))->delete($medium->path.$medium->medium_name);
+            }
 
             $medium->delete();
         }
